Switching length type from usize to u32
When trying to deploy the contract, it gives this error when the type of length is usize:

Error: rpc error: code = Unknown desc = rpc error: code = Unknown desc = failed to execute message; message index: 0: Error calling the VM: Error compiling Wasm: Could not compile: WebAssembly translation error: Error in middleware Gatekeeper: Float operator detected: F64Load { memarg: MemoryImmediate { align: 3, offset: 8, memory: 0 } }. The use of floats is not supported.: create wasm contract failed [CosmWasm/wasmd@v0.33.0/x/wasm/keeper/keeper.go:160] With gas wanted: '0' and gas used: '9525292' : unknown request

+++ contracts/voting-verifier/src/events.rs
@@ -187,7 +187,7 @@ fn parse_message_id(
             Ok((id.tx_hash_as_hex(), 0))
         }
         MessageIdFormat::Bech32m { prefix, length } => {
-            let bech32m_message_id = Bech32mFormat::from_str(prefix, *length, message_id)
+            let bech32m_message_id = Bech32mFormat::from_str(prefix, *length as usize, message_id)
                 .map_err(|_| ContractError::InvalidMessageID(message_id.into()))?;
             Ok((bech32m_message_id.to_string().try_into()?, 0))
         }
